Fixed the stupid arrow gap problem in classic, and multiple end points not having an associtated grahpcic (flag) in golf.
git-svn-id: svn://10.0.50.55/puffbomb/trunk@2114 d14347b8-48fd-0310-9e84-97907d434d85

+++ Code/Game/PuffBOMB/ClassicGame/ClassicGameEngine.cpp
@@ -333,11 +333,11 @@ void cClassicGameEngine::Draw() {
 
 	// Draw Points of interest tracking //
 	for ( size_t idx = 0; idx < PointsOfInterest.size(); idx++ ) {
-		ElementTracker( ArrowAnimator, PointsOfInterest[ idx ]->BoundingRect.ToRect(), true, -Real( 64 ) );
+		ElementTracker( ArrowAnimator, PointsOfInterest[ idx ]->BoundingRect.ToRect(), true, -Real( 48 ) );
 		
 		if( PointsOfInterestAnimators.size() == PointsOfInterest.size() )
 		{
-			ElementTracker( PointsOfInterestAnimators[ idx ], PointsOfInterest[ idx ]->BoundingRect.ToRect(), false, Real( 40 ) );
+			ElementTracker( PointsOfInterestAnimators[ idx ], PointsOfInterest[ idx ]->BoundingRect.ToRect(), false, Real( 32 ) );
 		}
 	}
